+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 This project uses [Gitmoji](https://gitmoji.carloscuesta.me/)
 
+## 6.0.2
+
+-   :heavy_minus_sign: remove `fast-deep-equal`.
+-   :sparkles: add in-house `deepEqual`.
+-   :arrow_up: dependency upgrade.
+
 ## 6.0.1
 
 -   :heavy_plus_sign: add `fast-deep-equal`.